Move `DomainStrategy` & `DialerProxy` to `DialSystem`
Fix https://github.com/XTLS/Xray-core/issues/608
